refactor logic to export to excel and pdf
