+++ model_report/exporters/base.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+class Exporter(object):
+
+    @classmethod
+    def render(cls, report, column_labels, report_rows, report_inlines):
+        raise NotImplementedError()
